move project information
