+++ src/pages/projects/[projectId]/summary/index.tsx
@@ -1,7 +1,7 @@
-import { Suspense } from "react"
+import { Suspense, useState } from "react"
 import { Routes } from "@blitzjs/next"
 import Link from "next/link"
-import { useQuery } from "@blitzjs/rpc"
+import { setQueryData, useMutation, useQuery } from "@blitzjs/rpc"
 import { useParam } from "@blitzjs/next"
 import Layout from "src/core/layouts/Layout"
 import getProjectData from "src/summary/queries/getProjectData"
@@ -10,19 +10,151 @@ import { MemberPrivileges } from "db"
 import DateFormat from "src/core/components/DateFormat"
 import DownloadJSON from "src/forms/components/DownloadJSON"
 import { MetadataDisplay } from "src/projects/components/MetaDataDisplay"
+import { JsonFormModal } from "src/core/components/JsonFormModal"
+import DownloadXLSX from "src/forms/components/DownloadXLSX"
+import getJsonSchema from "src/forms/utils/getJsonSchema"
+import toast from "react-hot-toast"
+import updateProject from "src/projects/mutations/updateProject"
+import { InformationCircleIcon } from "@heroicons/react/24/outline"
+import { Tooltip } from "react-tooltip"
 
 const Summary = () => {
   // Get data
   // Get projects
   const projectId = useParam("projectId", "number")
   const [project] = useQuery(getProjectData, { id: projectId })
+  const [updateProjectMutation] = useMutation(updateProject)
+
+  // State to store metadata
+  const [assignmentMetadata, setAssignmentMetadata] = useState(project.metadata)
+
+  const handleJsonFormSubmit = async (data) => {
+    //console.log("Submitting form data:", data) // Debug log
+    try {
+      const updatedProject = await updateProjectMutation({
+        id: project.id,
+        name: project.name,
+        metadata: data.formData,
+      })
+
+      // Update local state
+      setAssignmentMetadata(data.formData)
+
+      // Update the query data to refresh the background
+      await setQueryData(getProjectData, { id: projectId }, (prevData) => {
+        if (!prevData) {
+          throw new Error("No previous data found")
+        }
+
+        return {
+          ...prevData,
+          metadata: updatedProject.metadata,
+          formVersion: prevData.formVersion ?? null, // Ensure formVersion is explicitly null if undefined
+        }
+      })
+
+      toast.success("Form data has been successfully saved!")
+    } catch (error) {
+      console.error("Failed to save form data:", error)
+      toast.error("Failed to save form data. Please try again.")
+    }
+  }
+
+  const handleJsonFormError = (errors) => {
+    console.log(errors)
+  }
+  // Handle reset metadata
+  // Using hard reset to bypass validation
+  const handleResetMetadata = async () => {
+    try {
+      // Reset the metadata to an empty object
+      await updateProjectMutation({
+        id: project.id,
+        name: project.name,
+        metadata: {}, // Reset metadata to an empty object
+      })
+
+      // Update local state
+      setAssignmentMetadata({})
+
+      // Show a success toast
+      toast.success("Metadata has been successfully reset!")
+
+      // Refresh the form data
+      await setQueryData(getProjectData, { id: projectId }, (prevData) => {
+        if (!prevData) {
+          throw new Error("No previous data found")
+        }
+
+        return {
+          ...prevData,
+          metadata: {}, // Reset metadata to an empty object
+          formVersion: prevData.formVersion ?? null, // Ensure formVersion is explicitly null if undefined
+        }
+      })
+    } catch (error) {
+      console.error("Failed to reset metadata:", error)
+      toast.error("Failed to reset metadata. Please try again.")
+    }
+  }
 
   return (
     <main className="flex flex-col mt-2 mx-auto w-full max-w-7xl">
-      <h1 className="flex justify-center mb-2 text-3xl">Project Summary</h1>
+      <h1 className="flex justify-center items-center mb-2 text-3xl">
+        Project Summary
+        <InformationCircleIcon
+          className="h-6 w-6 ml-2 text-info stroke-2"
+          data-tooltip-id="project-download-tooltip"
+        />
+        <Tooltip
+          id="project-download-tooltip"
+          content="You can download information about just the project metadata, all the project data, or launch the project summary interactive viewer."
+          className="z-[1099] ourtooltips"
+        />
+      </h1>
+      {/* buttons */}
+      <div className="flex flex-row justify-center mb-2">
+        {project.metadata && (
+          <>
+            <DownloadJSON
+              data={project.metadata}
+              fileName={project.name}
+              className="btn btn-primary"
+              type="button"
+              label="Project Metadata JSON"
+            />
+            <DownloadXLSX
+              data={project.metadata}
+              fileName={project.name}
+              className="btn btn-secondary mx-2"
+              type="button"
+              label="Project Metadata XSLX"
+            />
+            <JsonFormModal
+              schema={getJsonSchema(project.formVersion?.schema)}
+              uiSchema={getJsonSchema(project.formVersion?.uiSchema)}
+              metadata={project.metadata}
+              label={"Edit Project Metadata"}
+              classNames="btn-info"
+              onSubmit={handleJsonFormSubmit}
+              onError={handleJsonFormError}
+              resetHandler={handleResetMetadata}
+              modalSize="w-11/12 max-w-5xl"
+            />
+          </>
+        )}
+
+        <DownloadJSON
+          data={project}
+          fileName={`${project.name}`}
+          className="mx-2 btn btn-primary"
+          label="Entire Project JSON"
+        />
+        <button className="btn btn-secondary">Launch Viewer (coming soon)</button>
+      </div>
 
       {/* Project  information */}
-      <div className="flex flex-row justify-center m-2">
+      <div className="flex flex-row justify-center">
         <div className="card bg-base-300 mx-2 w-full">
           <div className="card-body">
             <div className="card-title">Project Metadata</div>
@@ -44,26 +176,8 @@ const Summary = () => {
                 className="btn btn-primary"
                 href={Routes.EditProjectPage({ projectId: projectId! })}
               >
-                Edit
+                Edit Project Settings
               </Link>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Metadata */}
-      <div className="flex flex-row justify-center m-2">
-        <div className="card bg-base-300 mx-2 w-full">
-          <div className="card-body">
-            <div className="card-title">Summary Stats</div>
-            <div className="w-full flex flex-row justify-center">
-              <DownloadJSON
-                data={project}
-                fileName={`${project.name}`}
-                className="btn btn-primary"
-              />
-
-              <button className="btn btn-secondary mx-2">Launch Viewer (coming soon)</button>
             </div>
           </div>
         </div>
@@ -86,3 +200,6 @@ const SummaryPage = () => {
 }
 
 export default SummaryPage
+function updateProjectMutation(arg0: { id: number; name: string; metadata: any }) {
+  throw new Error("Function not implemented.")
+}
