+++ src/summary/queries/getProjectData.ts
@@ -12,6 +12,7 @@ export default resolver.pipe(resolver.zod(GetProjectData), resolver.authorize(),
   const project = await db.project.findFirst({
     where: { id },
     include: {
+      formVersion: true,
       tasks: {
         include: {
           element: true,
